requirements.txt
attrs==22.1.0
bidict==0.22.0
cachelib==0.9.0
cffi==1.15.1
click==8.1.3
cryptography==38.0.4
exceptiongroup==1.0.4
Flask==2.2.2
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SocketIO==5.3.2
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
importlib-metadata==5.1.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
pluggy==1.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyngrok==5.2.1
pyOpenSSL==22.1.0
pyparsing==3.0.9
pytest==7.2.0
python-engineio==4.3.4
python-ldap==3.4.3
python-socketio==5.7.2
PyYAML==6.0
SQLAlchemy==1.4.44
tomli==2.0.1
Werkzeug==2.2.2
zipp==3.11.0
